move get admin logic to page container
