+++ UI/src/components/features/RipaPageContainer.vue
@@ -5,16 +5,13 @@
     :dark="isDark"
     :on-update-dark="handleUpdateDark"
   >
-    <!-- <v-alert type="success" dismissible>
-      Cache was cleared and updated at 2021-04-23 11:34:26.
-    </v-alert> -->
     <slot></slot>
   </ripa-page-wrapper>
 </template>
 
 <script>
 import RipaPageWrapper from '@/components/organisms/RipaPageWrapper'
-import { mapState, mapGetters } from 'vuex'
+import { mapState, mapGetters, mapActions } from 'vuex'
 
 export default {
   name: 'ripa-page-container',
@@ -31,10 +28,35 @@ export default {
 
   computed: {
     ...mapState(['isAdmin']),
-    ...mapGetters(['isOnline']),
+    ...mapGetters([
+      'isOnline',
+      'mappedBeats',
+      'mappedCities',
+      'mappedSchools',
+      'mappedStatutes',
+    ]),
   },
 
   methods: {
+    ...mapActions([
+      'checkCache',
+      'getBeats',
+      'getCities',
+      'getSchools',
+      'getStatutes',
+    ]),
+
+    async getAdminData() {
+      this.loading = true
+      await Promise.all([
+        this.getBeats(),
+        this.getCities(),
+        this.getSchools(),
+        this.getStatutes(),
+      ])
+      this.loading = false
+    },
+
     getDarkFromLocalStorage() {
       const value = localStorage.getItem('ripa_dark_theme')
       if (value === null) {
@@ -51,6 +73,11 @@ export default {
     setDarkToLocalStorage() {
       localStorage.setItem('ripa_dark_theme', +this.isDark)
     },
+  },
+
+  created() {
+    this.checkCache()
+    this.getAdminData()
   },
 }
 </script>
